requirements.txt
# UI
PyQt5

# LLM
anthropic
langchain-anthropic

# Browser mode
browser-use
playwright

# Desktop mode
mss
pyautogui
pywinauto
pillow

# MCP server (optional but recommended -- exposes desktop tools to Claude / IDEs)
mcp
